+++ js/dateRangePicker.js
@@ -19,7 +19,7 @@ document.addEventListener('DOMContentLoaded', () => {
                 if (!self.HTMLInputElement) return;
                 if (self.selectedDates[1]) {
                     self.selectedDates.sort((a, b) => +new Date(a) - +new Date(b));
-                    self.HTMLInputElement.value = `${self.selectedDates[0]} to ${self.selectedDates[self.selectedDates.length - 1]}`;
+                    self.HTMLInputElement.value = `${self.selectedDates[0]} - ${self.selectedDates[self.selectedDates.length - 1]}`;
                 } else if (self.selectedDates[0]) {
                     self.HTMLInputElement.value = self.selectedDates[0];
                 } else {
@@ -36,9 +36,23 @@ document.addEventListener('DOMContentLoaded', () => {
 
             var newOptions = options;
 
+            // Get specific pre-filled items for that datepicker
             if(dateRangePickerDiv.dataset.multiple) { 
                 newOptions.type = 'multiple';
             }
+
+            // See if it already has a value!
+            if(dateRangePickerDiv.value) {
+                // Get the dates
+                var dates = dateRangePickerDiv.value.split(" - ");
+
+                newOptions.settings.selected = {
+                    dates: [dates[0] + ":" + dates[1]],
+                    month: dates[0].split("-")[1] - 1,
+                    year: dates[0].split("-")[0],
+                }
+            }
+            
 
             // Create the calendar
             const dateRangePicker = new VanillaCalendar('#' + dateRangePickerDiv.id, options);