+++ templates/base.php
@@ -100,14 +100,14 @@ $resultsNum = checkResultCapacity();
                         break;
                 }?>
                 <div class="notice <?= $severity;?>">
-                    <div class="rounded-full group transition-all duration-500 hover:scale-110 p-0.5 bg-black shadow-xl shadow-dark/50 absolute top-0 right-0 translate-x-1/4 -translate-y-1/4">
-                        <div class="w-full h-full rounded-full p-1 bg-white flex justify-center items-center">
-                            <svg class="h-4 w-4" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 384 512">
+                    <div class="notice-content-wrapper">
+                        <div class="notice-content">
+                            <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 384 512">
                                 <path d="M342.6 150.6c12.5-12.5 12.5-32.8 0-45.3s-32.8-12.5-45.3 0L192 210.7 86.6 105.4c-12.5-12.5-32.8-12.5-45.3 0s-12.5 32.8 0 45.3L146.7 256 41.4 361.4c-12.5 12.5-12.5 32.8 0 45.3s32.8 12.5 45.3 0L192 301.3 297.4 406.6c12.5 12.5 32.8 12.5 45.3 0s12.5-32.8 0-45.3L237.3 256 342.6 150.6z"/>
                             </svg>
                         </div>
                     </div>
-                    <p class="text-white"><?= $notice[0];?></p>
+                    <p><?= $notice[0];?></p>
                 </div>
             <?php endforeach;?>
         </div>