+++ views/backup.php
@@ -23,16 +23,50 @@ $qcSettings = array_intersect_key($settings, array_flip($qcKeys));
 <link href="/js/vanilla-calendar/build/vanilla-calendar.min.css" rel="stylesheet">
 <script src="/js/vanilla-calendar/build/vanilla-calendar.min.js" defer></script>
 <script src="/js/dateRangePicker.js" defer></script>
+<script src="/js/exportCheck.js" defer></script>
 
-<p class="text-center">Here's an example of a calendar.... nice!</p>
-
-<form class="form">
+<form class="form" method="POST" action="/process">
+    <input type="hidden" name="action" value="export-results">
+    <input type="hidden" name="return-path" value="<?= $currentURL;?>">
     <div class="form-fields">
-        <div class="input-wrapper max-w-sm">
-            <svg class="cursor-pointer" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 448 512">
-                <path d="M128 0c17.7 0 32 14.3 32 32V64H288V32c0-17.7 14.3-32 32-32s32 14.3 32 32V64h48c26.5 0 48 21.5 48 48v48H0V112C0 85.5 21.5 64 48 64H96V32c0-17.7 14.3-32 32-32zM0 192H448V464c0 26.5-21.5 48-48 48H48c-26.5 0-48-21.5-48-48V192zm64 80v32c0 8.8 7.2 16 16 16h32c8.8 0 16-7.2 16-16V272c0-8.8-7.2-16-16-16H80c-8.8 0-16 7.2-16 16zm128 0v32c0 8.8 7.2 16 16 16h32c8.8 0 16-7.2 16-16V272c0-8.8-7.2-16-16-16H208c-8.8 0-16 7.2-16 16zm144-16c-8.8 0-16 7.2-16 16v32c0 8.8 7.2 16 16 16h32c8.8 0 16-7.2 16-16V272c0-8.8-7.2-16-16-16H336zM64 400v32c0 8.8 7.2 16 16 16h32c8.8 0 16-7.2 16-16V400c0-8.8-7.2-16-16-16H80c-8.8 0-16 7.2-16 16zm144-16c-8.8 0-16 7.2-16 16v32c0 8.8 7.2 16 16 16h32c8.8 0 16-7.2 16-16V400c0-8.8-7.2-16-16-16H208zm112 16v32c0 8.8 7.2 16 16 16h32c8.8 0 16-7.2 16-16V400c0-8.8-7.2-16-16-16H336c-8.8 0-16 7.2-16 16z"/>
-            </svg>
-            <input id="dateRangePicker" class="date-range-picker" data-multiple="true" type="text" placeholder="Select Dates" readonly/>
+        <div class="field max-w-sm">
+            <label for="dateRangePicker">Select dates to backup/delete</label>
+            <div class="input-wrapper">
+                <svg class="cursor-pointer" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 448 512">
+                    <path d="M128 0c17.7 0 32 14.3 32 32V64H288V32c0-17.7 14.3-32 32-32s32 14.3 32 32V64h48c26.5 0 48 21.5 48 48v48H0V112C0 85.5 21.5 64 48 64H96V32c0-17.7 14.3-32 32-32zM0 192H448V464c0 26.5-21.5 48-48 48H48c-26.5 0-48-21.5-48-48V192zm64 80v32c0 8.8 7.2 16 16 16h32c8.8 0 16-7.2 16-16V272c0-8.8-7.2-16-16-16H80c-8.8 0-16 7.2-16 16zm128 0v32c0 8.8 7.2 16 16 16h32c8.8 0 16-7.2 16-16V272c0-8.8-7.2-16-16-16H208c-8.8 0-16 7.2-16 16zm144-16c-8.8 0-16 7.2-16 16v32c0 8.8 7.2 16 16 16h32c8.8 0 16-7.2 16-16V272c0-8.8-7.2-16-16-16H336zM64 400v32c0 8.8 7.2 16 16 16h32c8.8 0 16-7.2 16-16V400c0-8.8-7.2-16-16-16H80c-8.8 0-16 7.2-16 16zm144-16c-8.8 0-16 7.2-16 16v32c0 8.8 7.2 16 16 16h32c8.8 0 16-7.2 16-16V400c0-8.8-7.2-16-16-16H208zm112 16v32c0 8.8 7.2 16 16 16h32c8.8 0 16-7.2 16-16V400c0-8.8-7.2-16-16-16H336c-8.8 0-16 7.2-16 16z"/>
+                </svg>
+                <input id="dateRangePicker" name="filterDates" class="date-range-picker" data-multiple="true" type="text" placeholder="Leave empty to select all results" readonly/>
+            </div>
         </div>
     </div>
+    <div class="flex items-center gap-2">
+        <div id="backupBtn" class="btn smaller-btn btn-green cursor-pointer">Backup</div>
+        <div id="deleteBtn" class="btn smaller-btn btn-red cursor-pointer">Delete</div>
+    </div>
 </form>
+
+<div id="genericModalWrapper">
+    <div class="overlay"></div>
+    <div id="deleteModal" class="generic-modal">
+        <div class="close-modal">
+            <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 512 512">
+                <path d="M256 512A256 256 0 1 0 256 0a256 256 0 1 0 0 512zM175 175c9.4-9.4 24.6-9.4 33.9 0l47 47 47-47c9.4-9.4 24.6-9.4 33.9 0s9.4 24.6 0 33.9l-47 47 47 47c9.4 9.4 9.4 24.6 0 33.9s-24.6 9.4-33.9 0l-47-47-47 47c-9.4 9.4-24.6 9.4-33.9 0s-9.4-24.6 0-33.9l47-47-47-47c-9.4-9.4-9.4-24.6 0-33.9z"/>
+            </svg>
+        </div>
+        <svg class="fill-red-500 h-10 w-auto mb-3" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 512 512">
+            <path d="M256 32c14.2 0 27.3 7.5 34.5 19.8l216 368c7.3 12.4 7.3 27.7 .2 40.1S486.3 480 472 480H40c-14.3 0-27.6-7.7-34.7-20.1s-7-27.8 .2-40.1l216-368C228.7 39.5 241.8 32 256 32zm0 128c-13.3 0-24 10.7-24 24V296c0 13.3 10.7 24 24 24s24-10.7 24-24V184c0-13.3-10.7-24-24-24zm32 224a32 32 0 1 0 -64 0 32 32 0 1 0 64 0z"/>
+        </svg>
+        <p class="text-center">Are you sure you want to delete <span id="datesToDelete" class="font-black text-red-500"></span> results?</p>
+        <form action="/process" method="POST">
+            <input type="hidden" name="action" value="delete-results">
+            <input type="hidden" name="return-path" value="<?= $currentURL;?>">
+            <input type="hidden" name="dateRange" class="hidden-date-range" value="">
+            <div class="w-full flex justify-center items-center gap-3 mt-3">
+                <button type="submit" class="btn smaller-btn trigger-loading">Yes</button>
+                <div class="cursor-pointer btn smaller-btn close-modal no-styles">Cancel</div>
+            </div>
+        </form>
+    </div>
+</div>
+
+<?php // TODO: Eventually, it may be worth having a log of backups/deletions.